# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.AddPublic:
    properties:
      address:
        maxLength: 500
        minLength: 3
        type: string
      category:
        type: string
      city_id:
        type: string
      city_name:
        type: string
      district_id:
        type: string
      district_name:
        type: string
      email:
        type: string
      employee_count:
        type: integer
      is_educated:
        type: boolean
      last_visit_at:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        maxLength: 200
        minLength: 3
        type: string
      phone:
        maxLength: 20
        minLength: 5
        type: string
      postal_code:
        type: string
      province_id:
        type: string
      province_name:
        type: string
      visit_count:
        type: integer
    required:
    - address
    - category
    - city_id
    - city_name
    - district_id
    - district_name
    - name
    - province_id
    - province_name
    type: object
...
  dto.UpdatePublic:
    properties:
      address:
        maxLength: 500
        minLength: 3
        type: string
      category:
        type: string
      city_id:
        type: string
      city_name:
        type: string
      district_id:
        type: string
      district_name:
        type: string
      email:
        type: string
      employee_count:
        type: integer
      is_educated:
        type: boolean
      last_visit_at:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        maxLength: 200
        minLength: 3
        type: string
      phone:
        maxLength: 20
        minLength: 5
        type: string
      postal_code:
        type: string
      province_id:
        type: string
      province_name:
        type: string
      visit_count:
        type: integer
    type: object
...
  response.Error:
    properties:
      code:
        type: integer
      data: {}
      error: {}
      log_id:
        type: string
      message:
        type: string
      status:
        type: boolean
    type: object
  response.Success:
    properties:
      code:
        type: integer
      data: {}
      error: {}
      log_id:
        type: string
      message:
        type: string
      status:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: zaiduszhuhur@gmail.com
    name: API Support
    url: https://www.linkedin.com/in/zaidus-zhuhur/
  description: This is a sample server for a safety riding service.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Safety Riding API
  version: "1.0"
paths:
...
  /public:
    post:
      consumes:
      - application/json
      description: Create a new public entity record
      parameters:
      - description: Public payload
        in: body
        name: public
        required: true
        schema:
          $ref: '#/definitions/dto.AddPublic'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - ApiKeyAuth: []
      summary: Create a public entity
      tags:
      - Publics
  /public/{id}:
    delete:
      consumes:
      - application/json
      description: Delete public entity by ID
      parameters:
      - description: Public entity ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - ApiKeyAuth: []
      summary: Delete a public entity
      tags:
      - Publics
    get:
      consumes:
      - application/json
      description: Retrieve public entity by ID
      parameters:
      - description: Public entity ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - ApiKeyAuth: []
      summary: Get public entity detail
      tags:
      - Publics
    put:
      consumes:
      - application/json
      description: Update public entity information by ID
      parameters:
      - description: Public entity ID
        in: path
        name: id
        required: true
        type: string
      - description: Public payload
        in: body
        name: public
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePublic'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - ApiKeyAuth: []
      summary: Update a public entity
      tags:
      - Publics
  /publics:
    get:
      consumes:
      - application/json
      description: Retrieve paginated public entities with optional filters
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: order
        type: string
      - description: Filter by district ID
        in: query
        name: district_id
        type: string
      - description: Filter by city ID
        in: query
        name: city_id
        type: string
      - description: Filter by province ID
        in: query
        name: province_id
        type: string
      - description: Filter by category
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - ApiKeyAuth: []
      summary: List public entities
      tags:
      - Publics
  /publics/education-stats:
    get:
      consumes:
      - application/json
      description: Retrieve aggregated education statistics for public entities
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by district ID
        in: query
        name: district_id
        type: string
      - description: Filter by city ID
        in: query
        name: city_id
        type: string
      - description: Filter by province ID
        in: query
        name: province_id
        type: string
      - description: Filter by category
        in: query
        name: category
        type: string
      - description: Filter by education flag
        in: query
        name: is_educated
        type: string
      - description: Filter by month
        in: query
        name: month
        type: string
      - description: Filter by year
        in: query
        name: year
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - ApiKeyAuth: []
      summary: Get public education statistics
      tags:
      - Publics
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"